+++ NEWS.md
@@ -1,6 +1,9 @@
-# Development version
+# levitate 0.2.0
 
+* New ranking functions `lev_score_multiple()` and `lev_best_match()` (#5)
 * New package logo added by [`@zatch3301`](https://github.com/zatch3301) (#6)
+* New `lev_weighted_*()` functions (#10)
+
 
 # levitate 0.1.0